requirements.txt
PyYAML>=6.0
openpyxl>=3.1
pypdf>=4.0
pypdfium2>=5.6
rapidocr_onnxruntime>=1.2
